otaclient.errors: not using api_types anymore, move out of app package
