+++ src/otaclient/boot_control/_jetson_uefi.py
@@ -33,7 +33,7 @@ from pydantic import BaseModel
 from typing_extensions import Self
 
 from otaclient._types import OTAStatus
-from otaclient.app import errors as ota_errors
+from otaclient import errors as ota_errors
 from otaclient.app.configs import config as cfg
 from otaclient.boot_control._firmware_package import (
     FirmwareManifest,
